Add files via upload
This mini project, titled "Student Report Card Generator," is developed using C++. 
It allows teachers or administrators to generate a report card for students based on their marks in different subjects. 
The system takes input for student details such as name, roll number, and marks in various subjects. 
It then calculates the total marks, percentage and assigns a grade automatically. 
This project demonstrates the use of object-oriented programming (OOP) concepts such as classes, objects, and functions in C++.